requirements.txt
# Exchange connectivity
ccxt>=4.0.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Technical indicators
ta>=0.11.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Plotting (optional, for equity curves)
matplotlib>=3.7.0

# Date handling
python-dateutil>=2.8.0

# Web dashboard
flask>=2.3.0
flask-cors>=4.0.0
